+++ src/storage/invertedindex/index_builder.cpp
@@ -1,0 +1,13 @@
+module;
+
+module index_builder;
+import stl;
+import third_party;
+
+namespace infinity {
+
+IndexBuilder::IndexBuilder() {}
+
+IndexBuilder::~IndexBuilder() {}
+
+} // namespace infinity
