+++ src/storage/invertedindex/index_builder.cppm
@@ -1,0 +1,13 @@
+module;
+
+export module index_builder;
+import stl;
+import third_party;
+
+namespace infinity {
+export class IndexBuilder {
+public:
+    IndexBuilder();
+    ~IndexBuilder();
+};
+} // namespace infinity
